Add note about pre-existing cluster workers
If an ipcluster worker is still running when a new cluster is started (could be caused by terminating the ipcluster process without ipcluster stop) then you may have python workers still running. And the suggested remedy is to kill your open python sessions (one of which would be the hanging worker).