Update test_SurveySimulation.py
Added specific functionality to the revisit_wait_check method. 
The SurveySim is initialized, therefore revisit_wait is checked if it is None or not, it is checked if it is numerically equal to the input and it has been converted into an attribute. Finally, assert the calculation in the "scheduleRevisit" method.